… esphome/components/api/api_pb2_size.h