## What does this implement/fix?

This PR optimizes protobuf nested message encoding by eliminating temporary buffer allocations and expensive memory operations, achieving **2.3x faster performance** for all nested message operations.

The primary change is in the `ProtoWriteBuffer::encode_message()` method, which is used whenever a protobuf message contains another message as a field (e.g., BLE advertisements containing individual advertisement messages). The remaining changes are code reorganization to resolve circular dependencies.

### Key Change:

**Before:** The `encode_message()` method created inefficient temporary buffers:
```cpp
void encode_message(uint32_t field_id, const ProtoMessage &value, bool force) {
  this->encode_field_raw(field_id, 2);
  size_t begin = this->buffer_->size();
  value.encode(*this);
  const uint32_t nested_length = this->buffer_->size() - begin;
  // Problem: Creates temporary vector and uses expensive insert operation
  std::vector<uint8_t> var;
  ProtoVarInt(nested_length).encode(var);
  this->buffer_->insert(this->buffer_->begin() + begin, var.begin(), var.end());
}
```

**After:** Direct encoding with pre-calculated size:
```cpp
void encode_message(uint32_t field_id, const ProtoMessage &value, bool force) {
  this->encode_field_raw(field_id, 2);  // type 2: Length-delimited message
  
  // Calculate the message size first
  uint32_t msg_length_bytes = 0;
  value.calculate_size(msg_length_bytes);
  
  // Calculate how many bytes the length varint needs
  uint32_t varint_length_bytes = ProtoSize::varint(msg_length_bytes);
  
  // Reserve exact space for the length varint
  size_t begin = this->buffer_->size();
  this->buffer_->resize(this->buffer_->size() + varint_length_bytes);
  
  // Write the length varint directly
  ProtoVarInt(msg_length_bytes).encode_to_buffer_unchecked(
    this->buffer_->data() + begin, varint_length_bytes);
  
  // Now encode the message content - it will append to the buffer
  value.encode(*this);
  
  // Verify that the encoded size matches what we calculated
  assert(this->buffer_->size() == begin + varint_length_bytes + msg_length_bytes);
}
```

### Additional Changes:
- Consolidated `api_pb2_size.h` into `proto.h` to resolve circular dependencies
- Updated includes in `api_pb2.h`, `api_pb2.cpp`, and `api_frame_helper.cpp`

### Performance Impact:

This optimization eliminates:
- Temporary `std::vector<uint8_t>` allocation for every nested message
- O(n) memory move operation from `insert()` when prepending the length
- Redundant encoding to determine message size

**Benchmark Results (5 BLE advertisements):**

Before optimization:
- Average: 254.045 µs total (50.809 µs per advertisement)
- Encode: 183.547 µs (72.2% of total)

After optimization:
- Average: 110.963 µs total (22.193 µs per advertisement) 
- Encode: 56.840 µs (51.2% of total)

**Improvements:**
- **69.0% reduction** in encoding time (183.547 → 56.840 µs)
- **56.3% faster** overall (254.045 → 110.963 µs)
- **2.3x speedup** per advertisement (50.809 → 22.193 µs)
- Much lower variance (std dev: 16.598 → 3.054 µs)

This optimization benefits any protobuf message containing nested messages, including:
- Bluetooth proxy advertisements
- Entity state updates with nested fields
- Service responses with complex data structures

## Types of changes

- [ ] Bugfix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [x] Other (Performance optimization)

**Related issue or feature (if applicable):** fixes <link to issue>

**Pull request in [esphome-docs](https://github.com/esphome/esphome-docs) with documentation (if applicable):** esphome/esphome-docs#<number>

## Test Environment

- [x] ESP32
- [x] ESP32 IDF
- [ ] ESP8266
- [ ] RP2040
- [ ] BK72xx
- [ ] RTL87xx

## Example entry for `config.yaml`:
<!--
  Supplying a configuration snippet, makes it easier for a maintainer to test
  your PR. Furthermore, for new integrations, it gives an impression of how
  the configuration would look like.
  Note: Remove this section if this PR does not have an example entry.
-->

```yaml
# Configuration for testing BLE performance
esp32:
  board: esp32-poe-iso
  framework:
    type: esp-idf

bluetooth_proxy:
  active: true

api:
  services:
    - service: benchmark_ble_encode
      then:
        - lambda: |-
            ESP_LOGI("ble_benchmark", "=== BLE Raw Advertisements Encode Benchmark ===");
            
            // Add required includes
            #include "esphome/components/api/api_pb2.h"
            using namespace esphome::api;
            
            ESP_LOGI("ble_benchmark", "Creating test data...");
            
            // Create a response with a few advertisements
            BluetoothLERawAdvertisementsResponse response;
            
            // Add 5 mock advertisements with realistic data
            for (int i = 0; i < 5; i++) {
                BluetoothLERawAdvertisement adv;
                
                // Mock MAC address (6 bytes as uint64)
                adv.address = 0x112233445566ULL + i;
                
                // Typical RSSI value
                adv.rssi = -60 - (i * 2);
                
                // Address type (0 = public, 1 = random)
                adv.address_type = i % 2;
                
                // Mock advertisement data (typical BLE packet size ~31 bytes)
                adv.data = std::string(31, 0xAA);  // Simple fill pattern
                
                response.advertisements.push_back(adv);
            }
            
            ESP_LOGI("ble_benchmark", "Created %d advertisements", response.advertisements.size());
            
            // Warm up
            std::vector<uint8_t> buffer;
            uint32_t msg_size = 0;
            response.calculate_size(msg_size);
            buffer.reserve(msg_size);
            ProtoWriteBuffer write_buffer(&buffer);
            response.encode(write_buffer);
            
            ESP_LOGI("ble_benchmark", "Warmup: calculated size=%d, actual size=%d bytes", msg_size, buffer.size());
            
            // Run multiple test rounds for statistical accuracy
            const int rounds = 5;
            const int iterations_per_round = 500;
            
            float total_times[rounds];
            float calc_times[rounds];
            float encode_times[rounds];
            
            ESP_LOGI("ble_benchmark", "Running %d rounds of %d iterations each...", rounds, iterations_per_round);
            
            for (int round = 0; round < rounds; round++) {
                // Small delay between rounds
                delay(50);
                
                // Benchmark the actual API flow: calculate_size + pre-allocate + encode
                uint32_t start_time = micros();
                
                for (int i = 0; i < iterations_per_round; i++) {
                    // Step 1: Calculate size
                    uint32_t size = 0;
                    response.calculate_size(size);
                    
                    // Step 2: Pre-allocate buffer
                    std::vector<uint8_t> buf;
                    buf.reserve(size);
                    
                    // Step 3: Encode
                    ProtoWriteBuffer wb(&buf);
                    response.encode(wb);
                }
                
                uint32_t end_time = micros();
                total_times[round] = (end_time - start_time) / (float)iterations_per_round;
                
                // Also benchmark individual components for analysis
                // Calculate size only
                start_time = micros();
                for (int i = 0; i < iterations_per_round; i++) {
                    uint32_t size = 0;
                    response.calculate_size(size);
                }
                end_time = micros();
                calc_times[round] = (end_time - start_time) / (float)iterations_per_round;
                
                // Encode only (with pre-sized buffer)
                buffer.clear();
                buffer.reserve(msg_size);
                start_time = micros();
                for (int i = 0; i < iterations_per_round; i++) {
                    buffer.clear();
                    ProtoWriteBuffer wb(&buffer);
                    response.encode(wb);
                }
                end_time = micros();
                encode_times[round] = (end_time - start_time) / (float)iterations_per_round;
                
                ESP_LOGI("ble_benchmark", "Round %d: total=%.3f µs (calc=%.3f µs, encode=%.3f µs)", 
                         round + 1, total_times[round], calc_times[round], encode_times[round]);
            }
            
            // Calculate statistics
            float total_sum = 0, calc_sum = 0, encode_sum = 0;
            float total_min = total_times[0], total_max = total_times[0];
            float calc_min = calc_times[0], calc_max = calc_times[0];
            float encode_min = encode_times[0], encode_max = encode_times[0];
            
            for (int i = 0; i < rounds; i++) {
                total_sum += total_times[i];
                calc_sum += calc_times[i];
                encode_sum += encode_times[i];
                
                if (total_times[i] < total_min) total_min = total_times[i];
                if (total_times[i] > total_max) total_max = total_times[i];
                if (calc_times[i] < calc_min) calc_min = calc_times[i];
                if (calc_times[i] > calc_max) calc_max = calc_times[i];
                if (encode_times[i] < encode_min) encode_min = encode_times[i];
                if (encode_times[i] > encode_max) encode_max = encode_times[i];
            }
            
            float total_avg = total_sum / rounds;
            float calc_avg = calc_sum / rounds;
            float encode_avg = encode_sum / rounds;
            
            // Calculate standard deviation
            float total_variance = 0, calc_variance = 0, encode_variance = 0;
            for (int i = 0; i < rounds; i++) {
                float total_diff = total_times[i] - total_avg;
                float calc_diff = calc_times[i] - calc_avg;
                float encode_diff = encode_times[i] - encode_avg;
                total_variance += total_diff * total_diff;
                calc_variance += calc_diff * calc_diff;
                encode_variance += encode_diff * encode_diff;
            }
            total_variance /= rounds;
            calc_variance /= rounds;
            encode_variance /= rounds;
            
            float total_stddev = sqrt(total_variance);
            float calc_stddev = sqrt(calc_variance);
            float encode_stddev = sqrt(encode_variance);
            
            ESP_LOGI("ble_benchmark", "");
            ESP_LOGI("ble_benchmark", "=== FINAL RESULTS ===");
            ESP_LOGI("ble_benchmark", "Message size: %d bytes", buffer.size());
            ESP_LOGI("ble_benchmark", "Advertisements: %d", response.advertisements.size());
            ESP_LOGI("ble_benchmark", "");
            ESP_LOGI("ble_benchmark", "Total API Flow (calculate_size + allocate + encode):");
            ESP_LOGI("ble_benchmark", "  Average: %.3f µs (%.3f µs per advertisement)", 
                     total_avg, total_avg / response.advertisements.size());
            ESP_LOGI("ble_benchmark", "  Min: %.3f µs", total_min);
            ESP_LOGI("ble_benchmark", "  Max: %.3f µs", total_max);
            ESP_LOGI("ble_benchmark", "  Std Dev: %.3f µs", total_stddev);
            ESP_LOGI("ble_benchmark", "");
            ESP_LOGI("ble_benchmark", "Component Breakdown:");
            ESP_LOGI("ble_benchmark", "  Calculate Size: %.3f µs (%.1f%% of total)", calc_avg, (calc_avg / total_avg) * 100);
            ESP_LOGI("ble_benchmark", "  Encode: %.3f µs (%.1f%% of total)", encode_avg, (encode_avg / total_avg) * 100);
            ESP_LOGI("ble_benchmark", "  Overhead: %.3f µs (%.1f%% of total)", 
                     total_avg - calc_avg - encode_avg, 
                     ((total_avg - calc_avg - encode_avg) / total_avg) * 100);
            ESP_LOGI("ble_benchmark", "");
            ESP_LOGI("ble_benchmark", "=== End BLE Benchmark ===");
```

## Checklist:
  - [x] The code change is tested and works locally.
  - [ ] Tests have been added to verify that the new code works (under `tests/` folder).

If user exposed functionality or configuration variables are added/changed:
  - [ ] Documentation added/updated in [esphome-docs](https://github.com/esphome/esphome-docs).