@bdraco 2025-07-12
DeviceInfo has a lot of these so integration tests handle coverage well