@bdraco 2025-07-12
~34% reduction in cpu time on the proxy