@bdraco 2025-07-12
so esp32_ble_tracker is overhead is really the alloc overhead for each adv.

We can get rid of that by using a fixed size array