@bdraco 2025-07-12
once I get rid of the allocs
```
[08:23:44][D][time:048]: Synchronized time: 2025-07-12 08:23:44
[08:24:10][I][runtime_stats:038]: Component Runtime Statistics
[08:24:10][I][runtime_stats:039]: Period stats (last 60000ms):
[08:24:10][I][runtime_stats:060]:   bluetooth_proxy: count=7249, avg=0.06ms, max=21ms, total=415ms
[08:24:10][I][runtime_stats:060]:   esp32_ble_tracker: count=6855, avg=0.05ms, max=11ms, total=326ms
[08:24:10][I][runtime_stats:060]:   api: count=6855, avg=0.04ms, max=10ms, total=269ms
[08:24:10][I][runtime_stats:060]:   esp32_ble: count=6855, avg=0.02ms, max=14ms, total=139ms
[08:24:10][I][runtime_stats:060]:   status_led: count=6855, avg=0.01ms, max=3ms, total=66ms
[08:24:10][I][runtime_stats:060]:   esphome.ota: count=6855, avg=0.01ms, max=2ms, total=56ms
[08:24:10][I][runtime_stats:060]:   logger: count=4, avg=1.50ms, max=3ms, total=6ms
[08:24:10][I][runtime_stats:060]:   homeassistant.time: count=2, avg=2.00ms, max=3ms, total=4ms
[08:24:10][I][runtime_stats:060]:   template.sensor: count=1, avg=2.00ms, max=2ms, total=2ms
[08:24:10][I][runtime_stats:060]:   uptime.sensor: count=1, avg=2.00ms, max=2ms, total=2ms
[08:24:10][I][runtime_stats:060]:   ethernet: count=1, avg=0.00ms, max=0ms, total=0ms
[08:24:10][I][runtime_stats:060]:   preferences: count=1, avg=0.00ms, max=0ms, total=0ms
[08:24:10][I][runtime_stats:066]: Total stats (since boot):
[08:24:10][I][runtime_stats:078]:   bluetooth_proxy: count=14091, avg=0.06ms, max=21ms, total=777ms
[08:24:10][I][runtime_stats:078]:   esp32_ble_tracker: count=13697, avg=0.05ms, max=11ms, total=676ms
[08:24:10][I][runtime_stats:078]:   api: count=13697, avg=0.04ms, max=11ms, total=558ms
[08:24:10][I][runtime_stats:078]:   esp32_ble: count=13697, avg=0.02ms, max=14ms, total=219ms
[08:24:10][I][runtime_stats:078]:   status_led: count=13697, avg=0.01ms, max=3ms, total=132ms
[08:24:10][I][runtime_stats:078]:   esphome.ota: count=13697, avg=0.01ms, max=2ms, total=106ms
[08:24:10][I][runtime_stats:078]:   logger: count=6, avg=4.00ms, max=18ms, total=24ms
[08:24:10][I][runtime_stats:078]:   homeassistant.time: count=4, avg=1.50ms, max=3ms, total=6ms
[08:24:10][I][runtime_stats:078]:   template.sensor: count=2, avg=2.50ms, max=3ms, total=5ms
[08:24:10][I][runtime_stats:078]:   uptime.sensor: count=2, avg=2.00ms, max=2ms, total=4ms
[08:24:10][I][runtime_stats:078]:   ethernet: count=2, avg=0.00ms, max=0ms, total=0ms
[08:24:10][I][runtime_stats:078]:   preferences: count=2, avg=0.00ms, max=0ms, total=0ms
[08:24:14][D][time:048]: Synchronized time: 2025-07-12 08:24:14
[08:24:14][D][time:048]: Synchronized time: 2025-07-12 08:24:14
```

Both are now fast .. Will have to do that in a new pr